+++ core/src/main/java/jvmram/metrics/MetricsFactory.java
@@ -7,7 +7,7 @@ import jvmram.model.metrics.Os;
 import java.util.Map;
 
 public interface MetricsFactory {
-    Map<MetricType, RamMetric> getOrCreateMetrics(long pid, Os os);
+    Map<MetricType, RamMetric> getOrCreateMetrics(Integer pid, Os os);
 
     static MetricsFactory getInstance() {
         return MetricsFactoryImpl.INSTANCE;
